+++ packages/components/src/components/form/select-wrapper/select-wrapper.tsx
@@ -43,7 +43,6 @@ export class SelectWrapper {
 
   private select: HTMLSelectElement;
   private options: NodeListOf<HTMLOptionElement>;
-  private optgroups: NodeListOf<HTMLOptGroupElement>;
   private fakeOptionListNode: HTMLDivElement;
   private fakeOptionHighlightedNode: HTMLDivElement;
   private selectObserver: MutationObserver;
@@ -72,11 +71,13 @@ export class SelectWrapper {
   }
 
   public componentDidUnload(): void {
-    this.selectObserver.disconnect();
-    this.select.removeEventListener('mousedown', this.handleMouseEvents.bind(this));
-    this.select.removeEventListener('keydown', this.handleKeyboardEvents.bind(this));
-    if (!this.isTouch && typeof document !== 'undefined') {
-      document.removeEventListener('mousedown', this.handleClickOutside.bind(this), false);
+    if (!this.isTouch) {
+      this.selectObserver.disconnect();
+      this.select.removeEventListener('mousedown', this.handleMouseEvents.bind(this));
+      this.select.removeEventListener('keydown', this.handleKeyboardEvents.bind(this));
+      if (typeof document !== 'undefined') {
+        document.removeEventListener('mousedown', this.handleClickOutside.bind(this), false);
+      }
     }
   }
 
@@ -321,7 +322,6 @@ export class SelectWrapper {
 
   private setOptionList = (): void => {
     this.options = this.select.querySelectorAll('option');
-    this.optgroups = this.select.querySelectorAll('optgroup');
     this.optionSelected = this.select.selectedIndex;
     this.setOptionsDisabled();
   };
@@ -356,7 +356,7 @@ export class SelectWrapper {
     const PrefixedTagNames = getPrefixedTagNames(this.host, ['p-icon']);
 
     return Array.from(this.options).map((option: HTMLOptionElement, key: number) => [
-      this.optgroups.length > 0 && option === option.parentNode.firstChild && (
+      (option.parentElement.tagName === 'OPTGROUP' && option.previousElementSibling === null) && (
         <span class={cx(prefix('select-wrapper__fake-optgroup-label'))} role="presentation">
           {option.closest('optgroup').label}
         </span>