+++ packages/components/CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Removed
 - Stop browser support for **IE11** and **EdgeHTML**
 
+#### Fixed
+- Mix of `Optgroups` and `Options` on same level in `Select Wrapper` component
+
 ### [1.5.2] - 2020-07-22
 
 #### Fixed